+++ spec/unit/installer/user_project_integrator_spec.rb
@@ -122,12 +122,12 @@ module Pod
 
       #-----------------------------------------------------------------------#
 
-      describe "Helpers" do
+      describe "Private Helpers" do
 
         it "uses the path of the workspace defined in the podfile" do
-          path = "a_path"
-          @podfile.workspace(path)
-          @integrator.send(:workspace_path).should == path + ".xcworkspace"
+          path = Pathname.new("a_path")
+          @podfile.workspace(path.to_s)
+          @integrator.send(:workspace_path).should == path.sub_ext(".xcworkspace")
         end
 
         it "names the workspace after the user project if needed" do